+++ src/application/randomizer/Randomizer.cpp
@@ -67,11 +67,7 @@ void Randomizer::OnItemReceived(const int64_t itemId) const
         _saveFileManager.SetEventFlags(item.eventFlags);
     else if (item.type == ItemEmblem)
     {
-        PrintDebug(" --- Emblem received, total before: %d\n", _itemRepository.GetEmblemCount());
-        PrintDebug(" --- Emblem received, total before: %d\n", _gameStatus.unlock.currentEmblems);
         _itemRepository.AddEmblem();
-        PrintDebug(" --- Emblem received, total now: %d\n", _itemRepository.GetEmblemCount());
-        PrintDebug(" --- Emblem received, total now: %d\n", _gameStatus.unlock.currentEmblems);
     }
     else if (item.type == ItemFiller)
     {